reset files
