--- context-api/src/A.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function A() {
-    return (
-        <div>
-            
-        </div>
-    )
-}
