--- context-api/src/App.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-import ContextContainer from "./Context/ContextContainer";
-
-export default function App() {
-  return (
-    <div>
-      <ContextContainer>
-        <A />
-        <B />
-      </ContextContainer>
-    </div>
-  );
-}
