--- context-api/src/B.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function B() {
-    return (
-        <div>
-            
-        </div>
-    )
-}
