--- context-api/src/index.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
-
-ReactDOM.render(
-    <App />
-  document.getElementById('root')
-);
